+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Allow changes in the remote write api endpoint secret.
+
 ## [4.26.0] - 2023-03-20
 
 ### Changed